add exceptions to invalid delete an update of AlgoRelationType

+++ org.planqk.atlas.core/src/main/java/org/planqk/atlas/core/services/AlgoRelationTypeServiceImpl.java
@@ -6,7 +6,12 @@ import java.util.Set;
 import java.util.UUID;
 
 import org.planqk.atlas.core.model.AlgoRelationType;
+import org.planqk.atlas.core.model.exceptions.NotFoundException;
+import org.planqk.atlas.core.model.exceptions.SqlConsistencyException;
 import org.planqk.atlas.core.repository.AlgoRelationTypeRepository;
+import org.planqk.atlas.core.repository.AlgorithmRelationRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -14,9 +19,12 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class AlgoRelationTypeServiceImpl implements AlgoRelationTypeService {
+
+	private static final Logger LOG = LoggerFactory.getLogger(AlgoRelationType.class);
 	
 	@Autowired
 	private AlgoRelationTypeRepository repo;
+	@Autowired AlgorithmRelationRepository algorithmRelationRepository;
 
 	@Override
 	public AlgoRelationType save(AlgoRelationType algoRelationType) {
@@ -24,7 +32,7 @@ public class AlgoRelationTypeServiceImpl implements AlgoRelationTypeService {
 	}
 
 	@Override
-	public AlgoRelationType update(UUID id, AlgoRelationType algoRelationType) {
+	public AlgoRelationType update(UUID id, AlgoRelationType algoRelationType) throws NotFoundException {
 		// Check for type in database
 		Optional<AlgoRelationType> typeOpt = findById(id);
 		// If Type exists
@@ -35,12 +43,16 @@ public class AlgoRelationTypeServiceImpl implements AlgoRelationTypeService {
 			// Reference database type to set
 			return save(persistedType);
 		}
-		// TODO: Exception handling
-		return null;
+		LOG.info("Trying to update AlgoRelationType which does not exist.");
+		throw new NotFoundException("Cannot update AlgoRelationType since it could not be found.");
 	}
 
 	@Override
-	public void delete(UUID id) {
+	public void delete(UUID id) throws SqlConsistencyException {
+		if (algorithmRelationRepository.countRelationsUsingRelationType(id) > 0) {
+			LOG.info("Trying to delete algoRelationType that is used in at least 1 algorithmRelation.");
+			throw new SqlConsistencyException("Cannot delete algoRelationType, since it is used by existing algorithmRelations!");
+		}
 		repo.deleteById(id);
 	}
